src/core/models/img2name/requirements.txt
fastapi
uvicorn
numpy
tensorflow
pillow